requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
